refactor: move wikiWorker to a folder
