+++ src/services/wiki/wikiWorker/preload.ts
@@ -1,0 +1,3 @@
+import { fixPath } from '@services/libs/fixPath';
+
+fixPath();
